Objeto BaseDatos con BdAccess y PostgreSql

+++ BaseDatos.cs
@@ -20,73 +20,129 @@ namespace TodoASql
 	/// <summary>
 	/// Description of BaseDatos.
 	/// </summary>
-	public class BaseDatos
+	public abstract class BaseDatos:IDisposable
 	{
-		public BaseDatos()
+		protected IDbConnection con;
+		protected IDbCommand cmd;
+		protected BaseDatos(IDbConnection con)
 		{
+			Assert.IsNotNull(con);
+			this.con=con;
+			cmd=this.con.CreateCommand();
 		}
-		public static ADOX.CatalogClass CrearMDB(string nombreArchivo){
-			ADOX.CatalogClass cat=new CatalogClass();
-			cat.Create("Provider=Microsoft.Jet.OLEDB.4.0;" +
-				   "Data Source="+nombreArchivo+";" +
-				   "Jet OLEDB:Engine Type=5");
-			return cat;
+		public IDataReader ExecuteReader(string sentencia){
+			Assert.IsNotNull(con);
+			IDbCommand cmd_local=con.CreateCommand();
+			cmd_local.CommandText=sentencia;
+			return cmd_local.ExecuteReader();
 		}
-		public static OleDbConnection abrirMDB(string nombreArchivo){
-			OleDbConnection ConexionABase = new System.Data.OleDb.OleDbConnection();
-			ConexionABase.ConnectionString = 
-				@"PROVIDER=Microsoft.Jet.OLEDB.4.0;Data Source="+nombreArchivo;
-			ConexionABase.Open();
-			return ConexionABase;
+		public object ExecuteScalar(string sentencia){
+			Assert.IsNotNull(cmd);
+			cmd.CommandText=sentencia;
+			return cmd.ExecuteScalar();
 		}
-		public static OleDbConnection abrirPostgreOleDb(string Servidor,string Base, string Usuario, string Clave){
-			OleDbConnection ConexionABase = new System.Data.OleDb.OleDbConnection();
-			ConexionABase.ConnectionString="Provider=PostgreSQL OLE DB Provider;Data Source="+Servidor+";" +
-				"Location="+Base+";User id="+Usuario+";Password="+Clave+";"; // "timeout=1000";
-			ConexionABase.Open();
-			return ConexionABase;
+		public int ExecuteNonQuery(string sentencia){
+			Assert.IsNotNull(cmd);
+			cmd.CommandText=sentencia;
+			return cmd.ExecuteNonQuery();
 		}
-		public static OdbcConnection abrirPostgre(string Servidor,string Base, string Usuario, string Clave){
-			OdbcConnection ConexionABase = new OdbcConnection();
-			ConexionABase.ConnectionString=@"DRIVER=PostgreSQL Unicode;UID=import2sql;PORT=5432;SERVER=127.0.0.1;DATABASE=import2sqlDB;PASSWORD=sqlimport";
-			ConexionABase.Open();
-			return ConexionABase;
+		public bool EliminarTablaSiExiste(string nombreTabla){
+			Assert.IsNotNull(cmd);
+			try{
+				cmd.CommandText="DROP TABLE "+StuffTabla(nombreTabla);
+				cmd.ExecuteNonQuery();
+				return true;
+			}catch(DbException ex){
+				if(ex.ErrorCode==ErrorCode_NoExisteTabla){
+					return false; // ok, no existía la tabla por eso salto la excepción
+				}
+				throw;
+			}
 		}
+		public string StuffValor(object valor){
+			if(valor==null){
+				return "null";
+			}
+			if(valor.GetType()==typeof(String)){
+				return "'"+Cadena.BuscarYReemplazar((string) valor,"'","''")+"'";
+			}else if(valor.GetType()==typeof(double)){
+				return ((double) valor).ToString(Cadena.FormatoPuntoDecimal);
+			}else if(valor.GetType()==typeof(DateTime)){
+				return StuffFecha((DateTime) valor);
+			}else{
+				return valor.ToString();
+			}
+		}
+		public void Close(){
+			con.Close();
+			con=null;
+			cmd=null;
+		}
+		public void Dispose(){
+			Close();
+		}
+		public abstract int ErrorCode_NoExisteTabla{ get; }
+		public abstract string StuffTabla(string nombreTabla);
+		public abstract string StuffFecha(DateTime fecha);
+		public virtual string StuffCampo(string nombreCampo){ return StuffTabla(nombreCampo); }
 	}
 	[TestFixture]
 	public class ProbarBaseDatos{
-		[Test]
-		public void CreacionMdb(){
-			string nombreArchivo="tempAccesABorrar.mdb";
-			Archivo.Borrar(nombreArchivo);
-			Assert.IsTrue(!Archivo.Existe(nombreArchivo),"no debería existir");
-			Catalog cat=BaseDatos.CrearMDB(nombreArchivo);
-			Assert.IsTrue(Archivo.Existe(nombreArchivo),"debería existir");
-			IDbConnection con=BaseDatos.abrirMDB(nombreArchivo);
-			IDbCommand cmd=con.CreateCommand();
-			cmd.CommandText="CREATE TABLE tablaexistente (texto varchar(100), numero integer)";
-			cmd.ExecuteNonQuery();
-			cmd.CommandText="INSERT INTO tablaexistente (texto, numero) VALUES ('uno',1)";
-			cmd.ExecuteNonQuery();
-			AuxEnTodasLasBases(con);
+		public static void ObjEnTodasLasBases(BaseDatos db){
+			ObjOperacionesSimples(db);
+			ObjUsarReceptor(db);
 		}
-		[Test]
-		public void ConexionPostgre(){
-			
-			System.Windows.Forms.Application.OleRequired();
-			IDbConnection con=BaseDatos.abrirPostgre("127.0.0.1","import2sqlDB","import2sql","sqlimport");
-			AuxEnTodasLasBases(con);
-			IDbCommand cmd=con.CreateCommand();
-			cmd.CommandText="SELECT 3";
-			Assert.AreEqual(3,cmd.ExecuteScalar());
-			cmd.CommandText="SELECT 3.14";
-			Assert.AreEqual(3.14,cmd.ExecuteScalar());
+		public static void ObjOperacionesSimples(BaseDatos db){
+			IDataReader rdr=db.ExecuteReader("SELECT * FROM tablaexistente");
+			rdr.Read();
+			Assert.AreEqual("uno",rdr["texto"]);
+			rdr.Close();
+			Assert.AreEqual(0.625,db.ExecuteScalar("SELECT 10.0/16 FROM tablaexistente WHERE numero=1"));
+			db.EliminarTablaSiExiste("nueva_tabla_prueba");
+			db.ExecuteNonQuery(@"CREATE TABLE nueva_tabla_prueba (
+				nombre varchar(100),
+				numero integer,
+				monto double precision,
+				fecha date,
+				primary key (nombre));
+			");
+			db.ExecuteNonQuery("INSERT INTO nueva_tabla_prueba (nombre) VALUES ('solo')");
+			rdr=db.ExecuteReader("SELECT * FROM nueva_tabla_prueba");
+			rdr.Read();
+			Assert.AreEqual("solo",rdr["nombre"]);
+			rdr.Close();
 		}
-		public void AuxEnTodasLasBases(IDbConnection con){
+		public static void ObjUsarReceptor(BaseDatos db){
+			ReceptorSql receptor=new ReceptorSql(db,"nueva_tabla_prueba");
+			using(InsertadorSql insertador=new InsertadorSql(receptor)){
+				insertador["nombre"]="toto";
+				insertador["numero"]=2;
+				insertador["monto"]=3.1416;
+				insertador["fecha"]=new DateTime(2001,12,20);
+				insertador.InsertarSiHayCampos();
+			}
+			using(InsertadorSql insertador=new InsertadorSql(receptor)){
+				insertador["nombre"]="tute";
+				insertador["numero"]=3;
+				insertador["monto"]=Math.Exp(1);
+				insertador["fecha"]=new DateTime(1969,5,6);
+				insertador.InsertarSiHayCampos();
+			}
+			IDataReader rdr=db.ExecuteReader("SELECT * FROM nueva_tabla_prueba ORDER BY nombre");
+			rdr.Read();
+			rdr.Read();
+			Assert.AreEqual(3.1416,rdr["monto"]);
+			Assert.AreEqual(new DateTime(2001,12,20),rdr["fecha"]);
+			rdr.Read();
+			Assert.AreEqual(Math.Exp(1),(double) rdr["monto"],0.00000000000001);
+			Assert.AreEqual(new DateTime(1969,5,6),rdr["fecha"]);
+		}
+		/*
+		public static void AuxEnTodasLasBases(IDbConnection con){
 			AuxOperacionesSimples(con);
-			AuxUsarReceptor(con);
+			// AuxUsarReceptor(con);
 		}
-		public void AuxOperacionesSimples(IDbConnection con){
+		public static void AuxOperacionesSimples(IDbConnection con){
 			IDbCommand cmd=con.CreateCommand();
 			cmd.CommandText="SELECT * FROM tablaexistente";	
 			IDataReader rdr=cmd.ExecuteReader();
@@ -98,15 +154,11 @@ namespace TodoASql
 			try{
 				cmd.CommandText="DROP TABLE nueva_tabla_prueba";
 				cmd.ExecuteNonQuery();
-				Assert.Ignore("se pudo DROPear la tabla, la vez anterior la prueba se interrumpió porque no "+
-				              "debería existir. Hay que correr la prueba de vuelta. Lo normal era que no exista");
 			}catch(DbException ex){
-				switch(ex.ErrorCode){
-					case BdAccess.ErrorCode_NoExisteTabla: break; // ok Error de MDB
-					case PostgreSql.ErrorCode_NoExisteTabla: break; // ok Error de Postgre
-					default:
-						Assert.Ignore("DROP TABLE cantó "+ex.ErrorCode+": "+ex.Message+" no es un error conocido");
-						break;
+				if(ex.ErrorCode==BdAccess.ErrorCode_NoExisteTabla ||
+				   ex.ErrorCode==PostgreSql.ErrorCode_NoExisteTabla){
+				}else{
+					Assert.Ignore("DROP TABLE cantó "+ex.ErrorCode+": "+ex.Message+" no es un error conocido");
 				}
 			}
 			cmd.CommandText=@"CREATE TABLE nueva_tabla_prueba (
@@ -124,11 +176,23 @@ namespace TodoASql
 			rdr.Read();
 			Assert.AreEqual("solo",rdr["nombre"]);
 			rdr.Close();
-			cmd.CommandText="DROP TABLE nueva_tabla_prueba";
-			cmd.ExecuteNonQuery();
 		}
-		public void AuxUsarReceptor(IDbConnection con){
-			
+		public static void AuxUsarReceptor(IDbConnection con){
+			ReceptorSql receptor=new ReceptorSql(con,"nueva_tabla_prueba");
+			InsertadorSql insertador=new InsertadorSql(receptor);
+			insertador["nombre"]="toto";
+			insertador["numero"]=2;
+			insertador["monto"]=3.1416;
+			insertador["fecha"]=new DateTime(2001,12,20);
+			insertador.InsertarSiHayCampos();
+			IDbCommand cmd=con.CreateCommand();
+			cmd.CommandText="SELECT * FROM nueva_tabla_prueba ORDER BY nombre";
+			IDataReader rdr=cmd.ExecuteReader();
+			rdr.Read();
+			rdr.Read();
+			Assert.AreEqual(3.1416,rdr["monto"]);
+			Assert.AreEqual(new DateTime(2001,12,20),rdr["fecha"]);
 		}
+		*/
 	}
 }
